Mask_RCNN/requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.13.1
keras==2.1.0
opencv-python
h5py
imgaug
#IPython[all]